管理者用のseviseプロジェクトのインターフェース、クラスへ
