+++ practice/src/main/java/com/example/demo/admin/controller/AdminController.java
@@ -8,19 +8,19 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
+import com.example.demo.admin.service.AdminService;
 import com.example.demo.contact.entity.Contact;
-import com.example.demo.contact.service.ContactService;
 
 @Controller
 public class AdminController {
-
+	
 	@Autowired
-	private ContactService contactService;
+	private AdminService adminService;
 	
 	// 一覧
 	@GetMapping("/admin/contacts")
 	public String showContactList(Model model) {
-		List<Contact> contactList = contactService.getAllContacts();
+		List<Contact> contactList = adminService.getAllContacts();
 		model.addAttribute("contactList", contactList);
 		return "admin/contactList";
 	}
@@ -28,14 +28,14 @@ public class AdminController {
 	// 詳細
 	@GetMapping("/admin/contacts/{id}")
 	public String showDetail(@PathVariable Long id, Model model) {
-		model.addAttribute(contactService.getContactById(id).orElse(null));
+		model.addAttribute(adminService.getContactById(id).orElse(null));
 		return "admin/contactDetail";
 	}
 	
 	// 編集
 	@GetMapping("/admin/contacts/{id}/edit")
 	public String editContact(@PathVariable Long id, Model model) {
-		model.addAttribute("contactForm", contactService.getEditContact(id));
+		model.addAttribute("contactForm", adminService.getEditContact(id));
 		return "admin/contactEdit";
 	}
 }
